refactor(web): split session-list.tsx into sub-components
session-list.tsx: 2097 -> 484 lines (-77%). Extracted into
web/src/components/session-list/:
- session-item.tsx (1084) - StatusIndicator, SessionCard, SessionInfoDialog,
  ResultTextModal, SessionItemContent, InactiveProjectItem
- session-context-menu.tsx (257) - SessionContextMenu, ProjectContextMenu,
  shared grouping actions
- cwd-group.tsx (158) - CwdSessionGroup, CwdNode, DismissAllEndedButton
- session-sorting.tsx (138) - SortableNode, NewGroupDropTarget, GroupNode

Fixed: useSortable import moved from @dnd-kit/core to @dnd-kit/sortable
(was working before via session-list.tsx having both imports).

Pure structural move. Drag-drop tree manipulation stays in session-list.tsx
(closure-bound to sessionOrder state). Zustand selectors untouched.
